refactor: start splitting out allDocs utilities
